added save button to action bar in Add Issue; removed overflow menu from SGA feed
